WIP: Implement a dedicated node for reports.
